some code cleanup. database working
