+++ routes/api/patientRecordsAPI.js
@@ -13,18 +13,4 @@ router
   .put(recordController.update)
   .delete(recordController.remove);
 
-const userController = require("../../controllers/userController");
-
-// Matches with "/api/illnessRecords"
-router.route("/")
-  .get(userController.findAll)
-  .post(userController.create);
-
-// Matches with "/api/illnessRecords/:id"
-router
-  .route("/:id")
-  .get(userController.findById)
-  .put(userController.update)
-  .delete(userController.remove);
-
 module.exports = router;
