+++ routes/api/usersAPI.js
@@ -1,16 +1,16 @@
-// const router = require("express").Router();
-// const userController = require("../../controllers/userController");
+const router = require("express").Router();
+const userController = require("../../controllers/userController");
 
-// // Matches with "/api/illnessRecords"
-// router.route("/")
-//   .get(userController.findAll)
-//   .post(userController.create);
+// Matches with "/api/illnessRecords"
+router.route("/")
+  .get(userController.findAll)
+  .post(userController.create);
 
-// // Matches with "/api/illnessRecords/:id"
-// router
-//   .route("/:id")
-//   .get(userController.findById)
-//   .put(userController.update)
-//   .delete(userController.remove);
+// Matches with "/api/illnessRecords/:id"
+router
+  .route("/:id")
+  .get(userController.findById)
+  .put(userController.update)
+  .delete(userController.remove);
 
-// module.exports = router;
+module.exports = router;
